# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DialaDoc APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/prescription-medications": {
            "get": {
                "tags": [
                    "PrescriptionMedication"
                ],
                "summary": "getPrescriptionMedicationList",
                "description": "Get all PrescriptionMedications",
                "operationId": "59e735aecddb386cab493bd115ba757b",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PrescriptionMedication"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PrescriptionMedication"
                ],
                "summary": "createPrescriptionMedication",
                "description": "Create PrescriptionMedication",
                "operationId": "bd04a115ce80dba606c7aec0af017c4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PrescriptionMedication"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PrescriptionMedication"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/prescription-medications/{id}": {
            "get": {
                "tags": [
                    "PrescriptionMedication"
                ],
                "summary": "getPrescriptionMedicationItem",
                "description": "Get PrescriptionMedication",
                "operationId": "80a9205e217ac52c30433437c9fc6ce9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PrescriptionMedication",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PrescriptionMedication"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PrescriptionMedication"
                ],
                "summary": "updatePrescriptionMedication",
                "description": "Update PrescriptionMedication",
                "operationId": "fc67666f78bb10d6f9f5ca23a32240b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PrescriptionMedication",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PrescriptionMedication"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PrescriptionMedication"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PrescriptionMedication"
                ],
                "summary": "deletePrescriptionMedication",
                "description": "Delete PrescriptionMedication",
                "operationId": "5bdf279e079f5f55a365a6f964acea8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PrescriptionMedication",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PrescriptionMedication": {
                "required": [],
                "properties": {
                    "dosage": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "frequency": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}